yys_item/requirements.txt
aiocache==0.11.1
aiocqhttp==1.4.3
aiofiles==0.8.0
anyio==3.6.2
APScheduler==3.9.1.post1
arrow==1.2.3
async-timeout==4.0.2
backports.zoneinfo==0.2.1
binaryornot==0.4.4
blinker==1.5
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
chevron==0.14.0
click==8.1.3
colorama==0.4.6
cookiecutter==1.7.3
DBUtils==3.0.2
Deprecated==1.2.13
elastic-transport==8.4.0
elasticsearch==7.0.1
fastapi==0.79.1
greenlet==2.0.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
html2image==2.0.1
httpcore==0.14.7
httptools==0.5.0
httpx==0.22.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
mysql-connector==2.2.9
nb-cli==0.6.7
nonebot-adapter-onebot==2.1.5
nonebot-plugin-apscheduler==0.2.0
nonebot-plugin-gocqhttp==0.6.2
nonebot-plugin-guild-patch==0.2.1
nonebot-plugin-htmlrender==0.2.0.1
nonebot-plugin-txt2img==0.1.2
nonebot2==2.0.0rc1
numpy==1.23.5
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
playwright==1.28.0
poyo==0.5.0
priority==2.0.0
prompt-toolkit==3.0.32
psutil==5.9.4
py-cpuinfo==8.0.0
pydantic==1.9.2
pyee==9.0.4
pyfiglet==0.8.post1
Pygments==2.13.0
pygtrie==2.5.0
pymdown-extensions==9.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-markdown-math==0.8
python-slugify==6.1.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
quart==0.18.3
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
text-unidecode==1.3
toml==0.10.2
tomlkit==0.10.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
zipp==3.10.0
